+++ autoloads/gameplay_controller.gd
@@ -6,8 +6,6 @@ var gameplay_node: Node2D
 var current_room_scene: PackedScene
 var current_room_instance: Room
 
-var selected_unit: Unit
-
 var ally_units: Array[Unit]
 var enemy_units: Array[Unit]
 var gamestate: Dictionary[Vector2i, Tile]
@@ -95,15 +93,10 @@ func position_to_cell(position: Vector2):
 #endregion
 
 
-#region unit selection
-func select_unit_by_cell(cell: Vector2i):
-	var unit = gamestate[cell].unit
-	if unit:
-		selected_unit = unit
-		EventBus.unit_selected.emit(selected_unit)
+#region gamestate
 
-func  deselect_unit():
-	selected_unit = null
-	EventBus.unit_deselected.emit()
+func get_unit_by_cell(cell):
+	return gamestate[cell].unit
+
 
 #endregion