+++ content/abilities/validators/validator.gd
@@ -4,6 +4,9 @@ class_name Validator
 func validate(origin: Vector2i, area: Array[Vector2i]):
 	var output: Array[Vector2i] = []
 	for cell in area:
-		if true:
+		if validate_cell(origin, cell):
 			output.append(cell)
 	return output
+
+func validate_cell(_origin: Vector2i, _target: Vector2i):
+	return true