+++ gameplay/gameplay.gd
@@ -1,7 +1,8 @@
 extends Node2D
 
-@onready var main_grid = %MainGrid
-@onready var cursor = %Cursor
+@onready var main_grid: MainGrid = %MainGrid
+@onready var cursor: Cursor = %Cursor
 @onready var unit_container = %UnitContainer
 @onready var room_container = %RoomContainer
- 
+
+@onready var area_indicator: OffsetGrid = %AreaIndicator