+++ gameplay2/level/level.gd
@@ -1,6 +1,0 @@
-extends Resource
-
-class_name Level
-
-@export var name = "level1"
-#@export var enemies = Array[xxxx]