+++ ui/scenes/gameplay_screen/finishbuttontemp.gd
@@ -1,4 +1,0 @@
-extends Button
-
-func _pressed() -> void:
-	GameplayController.end_mission_gameplay()